Remove old ups script, move remaining functionalities to the emissions script
