+++ supply-chain/.env.SAMPLE
@@ -18,6 +18,4 @@ LEDGER_EMISSION_TOKEN_CONTRACT_ADDRESS="0xXXXXX"
 KEYCHAIN_NOLOG=Y
 # for IPFS connector
 IPFS_URL=http://127.0.0.1:5001
-# abort on any error from UPS by default, else set this to Y to skip packages with errors
-UPS_SKIP_ERRORS=N
